Warnings cleanup.

+++ libmesh/plugins/ply/PlyModelSystem.cpp
@@ -71,11 +71,11 @@ Mesh::VisualPtr PlyModelSystem::load(const Transfer::RemoteFileMetadata& metadat
 	ss >> s;
 
 	Mesh::MeshdataPtr mdp(new Mesh::Meshdata);
-	
+
 	int state = FILLER;
 	int vertexNum = 0, faceNum = 0, edgeNum = 0;
 	std::vector<int> propV, propF, propE;
-	
+
 	bool loop = true;
 	std::vector<String> file;
 	while(ss && loop) {
@@ -157,7 +157,7 @@ Mesh::VisualPtr PlyModelSystem::load(const Transfer::RemoteFileMetadata& metadat
 			int numTexCoords;
 			//obtain vertex data
 			for(int i = 0; i < vertexNum; i++) {
-				for(int j = 0; j < propV.size(); j++) {
+				for(uint32 j = 0; j < propV.size(); j++) {
 					ss >> temp;
 					switch(propV[j]) {
 						case X: vert[i][0] = temp; break;
@@ -174,7 +174,7 @@ Mesh::VisualPtr PlyModelSystem::load(const Transfer::RemoteFileMetadata& metadat
 			}
 			//obtain face data
 			for(int i = 0; i < faceNum; i++) {
-				for(int j = 0; j < propF.size(); j++) {
+				for(uint32 j = 0; j < propF.size(); j++) {
 					switch(propF[j]) {
 						case VI:
 							ss >> numIndices;
@@ -195,7 +195,7 @@ Mesh::VisualPtr PlyModelSystem::load(const Transfer::RemoteFileMetadata& metadat
 								ss >> tc[face[i][k]][0];
 								temp2 = tc[face[i][k]][1];
 								ss >> tc[face[i][k]][1];
-								
+
 								if(file.size() == 1) {
 									if((tc[face[i][k]][0] != temp1 && temp1 != -1) &&
 										(tc[face[i][k]][1] != temp2 && temp2 != -1))
@@ -203,7 +203,7 @@ Mesh::VisualPtr PlyModelSystem::load(const Transfer::RemoteFileMetadata& metadat
 									else
 										tc[face[i][k]][2] = 0;
 								}
-								
+
 							}
 							if(file.size() > 1) {
 								int numTex;
@@ -228,7 +228,7 @@ Mesh::VisualPtr PlyModelSystem::load(const Transfer::RemoteFileMetadata& metadat
 			nod.parent = Mesh::NullNodeIndex;
 			mdp->nodes.push_back(nod);
 			mdp->rootNodes.push_back(0);
-			
+
 			Mesh::SubMeshGeometry smg;
 
 			//make a primitive
@@ -241,7 +241,7 @@ Mesh::VisualPtr PlyModelSystem::load(const Transfer::RemoteFileMetadata& metadat
 			Mesh::GeometryInstance gi;
 			gi.geometryIndex = 0;
 			gi.parentNode = 0;
-			
+
 			//color: currently just takes the average of the colors and sets
 			//the color of the figure to this average color
 			int sumRed = 0, sumGreen = 0, sumBlue = 0, sumAlpha = 0;
@@ -269,7 +269,7 @@ Mesh::VisualPtr PlyModelSystem::load(const Transfer::RemoteFileMetadata& metadat
 							mdp->geometry[0].primitives[0].indices.push_back(face[i][j]);
 				} else {
 					//for each next set of indices, we first see
-					//if a previous primitive has either shared an 
+					//if a previous primitive has either shared an
 					//index or shared a point. If yes, we may add
 					//to the same primitive, or we may make a new
 					//submeshgeometry.
@@ -281,30 +281,30 @@ Mesh::VisualPtr PlyModelSystem::load(const Transfer::RemoteFileMetadata& metadat
 						int counterSMG = 0;
 						std::vector<int> hitSMG;
 						bool sameSMG = false;
-						while(counterSMG < mdp->geometry.size()) {
+						while(counterSMG < (int)mdp->geometry.size()) {
 							for(int j = 0; j < 3; j++) {
 								//iterate through the indices
-								for(int k = 0; k < mdp->geometry[counterSMG].primitives.size(); k++) {
-									for(int l = 0; l < mdp->geometry[counterSMG].primitives[k].indices.size(); l++) {
+								for(uint32 k = 0; k < mdp->geometry[counterSMG].primitives.size(); k++) {
+									for(uint32 l = 0; l < mdp->geometry[counterSMG].primitives[k].indices.size(); l++) {
 										//we must check if the index is the same and also if the texture is the same
 										//something is wrong here! reverseMap is creating new entries (and that shouldn't happen!)
 										if(reverseMap[counterSMG][mdp->geometry[counterSMG].primitives[k].indices[l]] == face[i][j] &&
 											((col[reverseMap[counterSMG][mdp->geometry[counterSMG].primitives[k].indices[l]]][1] == col[face[i][j]][1] &&
 											col[reverseMap[counterSMG][mdp->geometry[counterSMG].primitives[k].indices[l]]][2] == col[face[i][j]][2] &&
 											col[reverseMap[counterSMG][mdp->geometry[counterSMG].primitives[k].indices[l]]][3] == col[face[i][j]][3]) ||
-											(abs(tc[reverseMap[counterSMG][mdp->geometry[counterSMG].primitives[k].indices[l]]][2]) > file.size() && 
-											abs(tc[face[i][j]][2]) > file.size() && file.size() > 0))) {
+                                                                                            (abs(tc[reverseMap[counterSMG][mdp->geometry[counterSMG].primitives[k].indices[l]]][2]) > (int)file.size() &&
+                                                                                                abs(tc[face[i][j]][2]) > (int)file.size() && file.size() > 0))) {
 												sameSMG = true;
 												bool isThere = false;
-												for(int m = 0; m < hitSMG.size(); m++)
+												for(uint32 m = 0; m < hitSMG.size(); m++)
 													if(hitSMG[m] == counterSMG) isThere = true; //is there a better way of doing this?
 												if(!isThere) hitSMG.push_back(counterSMG);
 
-												
+
 										}
 									}
 								}
-								
+
 							}
 							counterSMG++;
 						}
@@ -313,13 +313,13 @@ Mesh::VisualPtr PlyModelSystem::load(const Transfer::RemoteFileMetadata& metadat
 							//if the point hit multiple SMG's, we will have to merge them before adding a point
 							while(hitSMG.size() > 1) {
 								//First, submeshgeometry
-								for(int j = 0; j < mdp->geometry[hitSMG[1]].primitives[0].indices.size(); j++) {//we loop through the indices of the smg to be destroyed
+								for(uint32 j = 0; j < mdp->geometry[hitSMG[1]].primitives[0].indices.size(); j++) {//we loop through the indices of the smg to be destroyed
 									int ind = reverseMap[hitSMG[0]].size();										//index for increasing the reverseMap size...
 									Vector3f point = Vector3f(vert[reverseMap[hitSMG[1]][mdp->geometry[hitSMG[1]].primitives[0].indices[j]]][0],				//here's a point from the smg to be transferred, at location j
 										vert[reverseMap[hitSMG[1]][mdp->geometry[hitSMG[1]].primitives[0].indices[j]]][1],
 										vert[reverseMap[hitSMG[1]][mdp->geometry[hitSMG[1]].primitives[0].indices[j]]][2]);
 									bool addPoint = true;
-									for(int k = 0; k < mdp->geometry[hitSMG[0]].positions.size() && addPoint; k++) {	//we loop through the other points in the smg to be stuffed
+									for(uint32 k = 0; k < mdp->geometry[hitSMG[0]].positions.size() && addPoint; k++) {	//we loop through the other points in the smg to be stuffed
 										if(point.x == mdp->geometry[hitSMG[0]].positions[k].x &&						//and try to see if there is a clone. if there is a clone,
 											point.y == mdp->geometry[hitSMG[0]].positions[k].y &&						//we don't add a point? but if there is, we do add a point?
 											point.z == mdp->geometry[hitSMG[0]].positions[k].z)
@@ -329,20 +329,20 @@ Mesh::VisualPtr PlyModelSystem::load(const Transfer::RemoteFileMetadata& metadat
 										mdp->geometry[hitSMG[0]].positions.push_back(point);					//since we need to add a point, we store the point in the smg...
 										indexMap[reverseMap[hitSMG[1]][mdp->geometry[hitSMG[1]].primitives[0].indices[j]]] = ind;								//we map the old point to the new point...
 										reverseMap[hitSMG[0]][ind] = reverseMap[hitSMG[1]][mdp->geometry[hitSMG[1]].primitives[0].indices[j]];					//and map the new point in reverse to the old!
-									} 
+									}
 									mdp->geometry[hitSMG[0]].primitives[0].indices.push_back(indexMap[reverseMap[hitSMG[1]][mdp->geometry[hitSMG[1]].primitives[0].indices[j]]]);
 								}
 								//proper deletions
 								mdp->geometry.erase(mdp->geometry.begin() + hitSMG[1]);
 								//then, geometryInstance
 								mdp->instances.erase(mdp->instances.begin() + hitSMG[1]);
-								for(int j = hitSMG[1]; j < mdp->instances.size(); j++)
+								for(uint32 j = hitSMG[1]; j < mdp->instances.size(); j++)
 									mdp->instances[j].geometryIndex = j;
 								//also, reverseMap
 								reverseMap.erase(reverseMap.begin() + hitSMG[1]);
 								//finally, hitSMG
 								hitSMG.erase(hitSMG.begin() + 1);
-								for(int i = 1; i < hitSMG.size(); i++) hitSMG[i]--;//since we removed a hitSMG we have to lower the index of the other hits by one
+								for(uint32 i = 1; i < hitSMG.size(); i++) hitSMG[i]--;//since we removed a hitSMG we have to lower the index of the other hits by one
 							}
 							for(int j = 0; j < 3; j++) {
 								//we put the point in
@@ -356,7 +356,7 @@ Mesh::VisualPtr PlyModelSystem::load(const Transfer::RemoteFileMetadata& metadat
 								}
 
 								mdp->geometry[hitSMG[0]].primitives[0].indices.push_back(indexMap[face[i][j]]);
-									
+
 							}
 						} else {
 							//otherwise, we'll make a new submeshgeometry/geometry set
@@ -378,18 +378,18 @@ Mesh::VisualPtr PlyModelSystem::load(const Transfer::RemoteFileMetadata& metadat
 						}
 					}
 				}
-				
+
 			}
 
 
 			//int fileCounter = 0;
 			int mapCounter = 0;
-			for(int i = 0; i < mdp->geometry.size(); i++) {
+			for(uint32 i = 0; i < mdp->geometry.size(); i++) {
 				//one texture set for the geometry
 				if(tc[0][0] > 0) {
 					Mesh::SubMeshGeometry::TextureSet ts;
 					ts.stride = 2;
-					for(int j = 0; j < mdp->geometry[i].positions.size(); j++) {
+					for(uint32 j = 0; j < mdp->geometry[i].positions.size(); j++) {
 						for(int k = 0; k < 2; k++)  {
 							ts.uvs.push_back(tc[reverseMap[i][j]][k]);
 						}
@@ -400,7 +400,7 @@ Mesh::VisualPtr PlyModelSystem::load(const Transfer::RemoteFileMetadata& metadat
 				if(faceNum > 0) {
 					Mesh::MaterialEffectInfo::Texture t;
 					bool hasTexture = false;
-					for(int j = 0; j < mdp->geometry[i].primitives[0].indices.size() && !hasTexture; j++) {
+					for(uint32 j = 0; j < mdp->geometry[i].primitives[0].indices.size() && !hasTexture; j++) {
 						if((abs(tc[reverseMap[i][mdp->geometry[i].primitives[0].indices[j]]][0]) > .001 &&
 							abs(tc[reverseMap[i][mdp->geometry[i].primitives[0].indices[j]]][0]) < INT_MAX) ||
 							(abs(tc[reverseMap[i][mdp->geometry[i].primitives[0].indices[j]]][1]) > .001 &&
@@ -410,13 +410,13 @@ Mesh::VisualPtr PlyModelSystem::load(const Transfer::RemoteFileMetadata& metadat
 					//if the texCoords are all almost zero (or are infinite), then this is not a truly texturized geometry
 
 					if(hasTexture && tc[reverseMap[i][mdp->geometry[i].primitives[0].indices[0]]][2] >= 0
-						&& tc[reverseMap[i][mdp->geometry[i].primitives[0].indices[0]]][2] < file.size()) {						
+						&& tc[reverseMap[i][mdp->geometry[i].primitives[0].indices[0]]][2] < file.size()) {
 
 						//we add the texture and typical values
 						mei.shininess = -128;
 						mei.reflectivity = -1;
 						bool newFile = true;
-						for(int l = 0; l < mdp->textures.size(); l++)
+						for(uint32 l = 0; l < mdp->textures.size(); l++)
 							if(mdp->textures[l] == file[tc[reverseMap[i][mdp->geometry[i].primitives[0].indices[0]]][2]])
 								newFile = false;
 						if(newFile)
@@ -438,7 +438,7 @@ Mesh::VisualPtr PlyModelSystem::load(const Transfer::RemoteFileMetadata& metadat
 					} else {
 						//we add a color
 						Vector4f c;
-						if(mdp->geometry[i].primitives[0].primitiveType == Mesh::SubMeshGeometry::Primitive::TRIANGLES && 
+						if(mdp->geometry[i].primitives[0].primitiveType == Mesh::SubMeshGeometry::Primitive::TRIANGLES &&
 							col[reverseMap[i][mdp->geometry[i].primitives[0].indices[0]]][0] >= 0) { //we might need a better check here if we get certain files
 							c = Vector4f(col[reverseMap[i][mdp->geometry[i].primitives[0].indices[0]]][0] / 255.0,
 								col[reverseMap[i][mdp->geometry[i].primitives[0].indices[0]]][1] / 255.0,
@@ -454,14 +454,14 @@ Mesh::VisualPtr PlyModelSystem::load(const Transfer::RemoteFileMetadata& metadat
 					mei.textures.push_back(t);
 					bool addMat = true;
 					int matPos = -1;
-					for(int j = 0; j < mdp->materials.size(); j++)
+					for(uint32 j = 0; j < mdp->materials.size(); j++)
 						if(mdp->materials[j] == mei) {
 							addMat = false;
 							matPos = j;
 						}
 					mdp->geometry[i].primitives[0].materialId = 1;
 					//if there is not, we add it, otherwise we do not
-					if(addMat) {		
+					if(addMat) {
 						mdp->instances[i].materialBindingMap[mdp->geometry[i].primitives[0].materialId] = mapCounter;
 						mapCounter++;
 						mdp->materials.push_back(mei);
@@ -470,7 +470,7 @@ Mesh::VisualPtr PlyModelSystem::load(const Transfer::RemoteFileMetadata& metadat
 
 				} else mdp->materials.push_back(mei);
 			}
-			
+
 			delete vert;
 			delete col;
 			delete face;
